refactor: LottoMachine 클래스 생성 및 메서드 이동
