+++ src/LottoGame.js
@@ -1,25 +1,18 @@
-const { Console, Random } = require("@woowacourse/mission-utils");
+const { Console } = require("@woowacourse/mission-utils");
 const { REQUEST_MESSAGE } = require("./constants/message.js");
-const {
-  LOTTO_PRICE,
-  LOTTO_NUM_MIN_RANGE,
-  LOTTO_NUM_MAX_RANGE,
-  LOTTO_DIGITS,
-  LOTTO_PRIZE_MATCH_COUNT,
-  LOTTO_PRIZE_MONEY,
-} = require("./constants/condition.js");
+const { LOTTO_PRIZE_MATCH_COUNT, LOTTO_PRIZE_MONEY } = require("./constants/condition.js");
 const LottoGameView = require("./LottoGameView.js");
-const Lotto = require("./Lotto.js");
 const Validation = require("./Validation.js");
+const LottoMachine = require("./LottoMachine.js");
 
 class LottoGame {
-  purchaseAmount;
   lottos;
   winningNumbers;
   bonusNumber;
 
   constructor() {
     this.LottoGameView = new LottoGameView();
+    this.lottoMachine = new LottoMachine();
   }
 
   play() {
@@ -28,28 +21,13 @@ class LottoGame {
 
   purchaseLottoPhase() {
     this.LottoGameView.requestInput(REQUEST_MESSAGE.PURCHASE_AMOUNT, (purchaseAmount) => {
-      Validation.validatePurchaseAmount(purchaseAmount);
-      this.purchaseAmount = Number(purchaseAmount);
-      const lottoQuantity = this.getLottoQuantity(this.purchaseAmount);
-      this.createLottos(lottoQuantity);
-      this.LottoGameView.printLottoQuantity(lottoQuantity);
+      this.lottoMachine.insertMoney(purchaseAmount);
+      this.lottos = this.lottoMachine.purchaseLottos();
+
+      this.LottoGameView.printLottoQuantity(this.lottos.length);
       this.LottoGameView.printEachLottoNumbers(this.lottos);
 
       this.saveWinningNumbersPhase();
-    });
-  }
-  getLottoQuantity(purchaseAmount) {
-    return purchaseAmount / LOTTO_PRICE;
-  }
-  generateLottoNumbers() {
-    return Random.pickUniqueNumbersInRange(LOTTO_NUM_MIN_RANGE, LOTTO_NUM_MAX_RANGE, LOTTO_DIGITS);
-  }
-  createLottos(lottoQuantity) {
-    this.lottos = Array.from({ length: lottoQuantity }, () => {
-      const lottoNumbers = this.generateLottoNumbers();
-      const ascendingNumbers = lottoNumbers.sort((numA, numB) => numA - numB);
-
-      return new Lotto(ascendingNumbers);
     });
   }
 
@@ -62,6 +40,7 @@ class LottoGame {
       this.saveBonusNumberPhase();
     });
   }
+
   saveBonusNumberPhase() {
     this.LottoGameView.requestInput(REQUEST_MESSAGE.BONUS_NUMBER, (bonusNumber) => {
       Validation.validateBonusNumber(bonusNumber, this.winningNumbers);
